+++ src/main/java/it/contactlab/hub/sdk/java/exceptions/HttpException.java
@@ -2,7 +2,7 @@ package it.contactlab.hub.sdk.java.exceptions;
 
 import com.mashape.unirest.http.exceptions.UnirestException;
 
-public class HttpException extends BaseException {
+public class HttpException extends ContactHubException {
 
   public HttpException(UnirestException exception) {
     super("The HTTP call failed with error: " + exception.getMessage(), exception);
